update GetMonthSummary to GetMonthSpending and reconstruct budgetSummary
