+++ components/GetMonthSpending.js
@@ -2,13 +2,11 @@ import React, { useEffect, useState } from 'react';
 import { collection, query, where, onSnapshot } from 'firebase/firestore';
 import { database,auth } from "../firebase/firebaseSetup";
 
-const GetMonthSummary = ({ month }) => {
+const GetMonthSpending = ({ month }) => {
   const [spending, setSpending] = useState(0);
-  const [budgetLimit, setBudgetLimit] = useState(0);
 
   useEffect(() => {
     const userUid = auth.currentUser.uid;
-
     const [year, monthString] = month.split('-');
     const monthDate = new Date(year, parseInt(monthString) - 1, 1);
     const monthStart = new Date(monthDate.getFullYear(), monthDate.getMonth(), 1);
@@ -28,12 +26,6 @@ const GetMonthSummary = ({ month }) => {
         where('date', '<=', monthEnd)
       );
 
-      // Query for Budgets fot the user
-      const budgetsQuery = query(
-        collection(database, 'Budgets'),
-        where('user', '==', userUid)
-      );
-
       // Listen for changes in Expenses
       const unsubscribeExpenses = onSnapshot(expensesQuery, (expenseSnapshot) => {
         let totalSpending = 0;
@@ -44,20 +36,10 @@ const GetMonthSummary = ({ month }) => {
         setSpending(totalSpending);
       });
 
-      // Listen for changes in Budgets
-      const unsubscribeBudgets = onSnapshot(budgetsQuery, (budgetSnapshot) => {
-        if (!budgetSnapshot.empty) {
-          const latestBudget = budgetSnapshot.docs[budgetSnapshot.docs.length - 1].data();
-          setBudgetLimit(latestBudget.limit || 0);
-        } else {
-          setBudgetLimit(0);
-        }
-      });
 
       return () => {
         // Unsubscribe from both snapshots when the component unmounts
         unsubscribeExpenses();
-        unsubscribeBudgets();
       };
     };
 
@@ -71,7 +53,7 @@ const GetMonthSummary = ({ month }) => {
   }, [month]);
 
   // Return spending and budgetLimit as an object
-  return { spending, budgetLimit };
+  return { spending };
 };
 
-export default GetMonthSummary;
+export default GetMonthSpending;
